+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved documentation to `docs/` folder
 - Flattened screenshot directory structure
 
+### Fixed
+- AutoPilot: CMP not displaying for strategy legs created from templates
+
 ## [0.5.0] - 2024-12-07
 
 ### Added